+++ client/src/views/Admin.vue
@@ -5,7 +5,8 @@ import Error from "../components/alerts/Error.vue"
 import AdminUsers from '../components/admin/AdminUsers.vue';
 import NewUserModal from "../components/admin/modals/NewUserModal.vue"
 import Success from '../components/alerts/Success.vue';
-import AdminCreateClient from '../components/admin/AdminCreateClient.vue';
+import AdminClients from '../components/admin/AdminClients.vue';
+import NewClientModal from '../components/admin/modals/NewClientModal.vue';
 
 const successMsg = ref()
 const auth = useAuthStore()
@@ -39,15 +40,15 @@ const setSucMsg = (successObj) => {
         </ul>
         <div class="tab-content" id="myTabContent">
             <div class="tab-pane fade show active" id="users-tab-pane" role="tabpanel" aria-labelledby="users-tab" tabindex="0">
-                <button type="button" class="btn btn-outline-success mt-3" data-bs-toggle="modal" data-bs-target="#exampleModal">
-                    Add User
-                </button>
+                <button type="button" class="btn btn-outline-success mt-3" data-bs-toggle="modal" data-bs-target="#exampleModal">Add User</button>
                 <AdminUsers @success-message="setSucMsg" @error-message="setErrMsg" :token="token" />
             </div>
             <div class="tab-pane fade" id="clients-tab-pane" role="tabpanel" aria-labelledby="clients-tab" tabindex="1">
-                <AdminCreateClient/>
+                <button type="button" class="btn btn-outline-success mt-3" data-bs-toggle="modal" data-bs-target="#newClientModal">Add Client</button>
+                <AdminClients/>
             </div>
         </div>
     </div>
     <NewUserModal @new-user-error-message="setErrMsg" :token="token"/>
+    <NewClientModal/>
 </template>
